Finished Week 7
I also tried a lot of thing to make VLD work, but I eventually found an alternative that shows me when I have leaks, but sadly not where. This is the best I'll get though. this is also why a lot of random files are "added "as I tried deleting certain folders etc.